+++ src/core/hw/tegra_x1/gpu/renderer/shader_decompiler/decoder/float_min_max.cpp
@@ -41,8 +41,9 @@ void EmitFmnmxC(DecoderContext& context, InstFmnmxC inst) {
         context, inst.base.pred, inst.base.pred_inv, inst.base.dst,
         inst.base.src_pred, inst.base.src_pred_inv, inst.base.src_a,
         inst.base.abs_a, inst.base.neg_a,
-        ir::Value::ConstMemory(CMem(inst.cbuf_slot, RZ, inst.cbuf_offset * 4),
-                               ir::ScalarType::F32),
+        ir::Value::ConstMemory(
+            CMem(inst.cbuf_slot, RZ, static_cast<u64>(inst.cbuf_offset * 4)),
+            ir::ScalarType::F32),
         inst.base.abs_b, inst.base.neg_b);
 }
